Changed store connection
Removed app container and replaced connection to Home. Refactor of tests to reflect changes with Home as a container. ApplicationLayout acts as a simple component now which displays children.
