+++ src/js/components/ApplicationLayout.js
@@ -6,7 +6,7 @@ export default class ApplicationLayout extends React.Component {
     return (
       <div>
         <Navbar />
-        {React.cloneElement(this.props.children, { ...this.props, key: undefined, ref: undefined })}
+        {this.props.children}
       </div>
     );
   }
