--- tests/containers/App.test.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import { shallow } from 'enzyme';
-import { expect } from 'chai';
-import App from '../../src/js/containers/App';
-
-describe('App Container', () => {
-
-  it('Waiting for Testing', () => {
-    return true;
-  });
-
-});
